Get indexing working with lazy loading.
This code only pulls in elements that match the "path" of the indexing. In other words, if asking for xml["elem"]["book"][1]["author"], we should *only* return elements that match 1) the names, 2) the hierarchy, and 3) the count. With indexing, we have to index enough elements until we reach the number requested.

I've introduced code to try to short circuit the checks for performance reasons - the checks spin all matched elements in backwards order and fail out as soon as a mismatch is found. One caveat, if there isn't a match, this will spin the entire XML document looking for the requested elements.