Add fluent configuration option.
This change introduces a new fluent configuration method as a way of
hopefully making it easy and straightforward to introduce new
configuration options in the future.

The first config option introduced in this commit is
"shouldProcessNamespaces" which is forwarded on to the NSXMLParser
instance (related to issue #30).

There is also a config option for lazy XML parsing which can be used
instead of calling "lazy" versus "parse".