Rename Error type
To simplify the do/catch statements, I’ve renamed the XMLIndexError to
simple Error. It is in the XMLIndexer namespace, which means that it
does not interfere with any other error type. Because of the renaming I
needed to change the legacy XMLIndexer.Error to .XMLError because
otherwise the compiler would complain about a redeclaration of “Error”.
I’ve also renamed the throwing equivalent of the subscripts from “key”
and “idx” to “byKey” and “byIndex”.

I did update the test, but i can’t compile it because I dont have Quick
and Nimble.

P.S. I’ve also changed Srings join method to the joinWithSeperator
method as changed in Xcode 7 b6