Add specs and support for mixed content
This commit adds initial support for mixed content (i.e. text + XML
inside an element) and should fix issue #33.

To support a variety of text/XML content, the general approach to text
had to change - before, text was handled via simple string
concatenation. The problem, though, is that this means that an element
either has children or text - no options.

Now, there is the concept of both TextElements and XMLElements - both
are stored as children of an element. This change should be API backwards
compatible.