XMLElement.text will never be nil
Per issue #132, XMLElement can't be nil because of how the reduce
implementation works. Instead, it will be an empty string. I think this
makes more sense than allowing for nils anyway.

A lot of the suggestions mentioned on issue #84 are related to reducing
the number of nil checks as well, so this should help alleviate some of
that.