Fix enum naming in type conversion, etc.
The recommended naming for enums is camel case so this commit changes
the XMLDeserializationError cases to be named properly. In addition,
Sequence was removed from XMLIndexer so-as to avoid a conflict between
XMLIndexer.element and Sequence.element in Swift 4 - enumerating over
results is still possible by using the `.all` method, though.