feat: bump to Swift 6, add swift-testing
I dropped old deprecation warnings as this is definitely a major version bump. For Swift 6 support, some of the types that were sent with errors are now wrapped in a "sendable" variety so that they're async safe.

The GitHub builds also heavily use matrix support for multiple platform builds now, too.